+++ app/src/main/java/com/example/anpu/circles/MainActivityFragment.java
@@ -8,6 +8,8 @@ import android.support.v4.app.FragmentActivity;
 import android.support.v4.view.ViewPager;
 import android.view.MenuItem;
 
+import com.example.anpu.circles.adapter.ViewPagerFragmentAdapter;
+
 import java.util.ArrayList;
 import java.util.List;
