+++ app/src/main/java/com/example/anpu/circles/adapter/ViewPagerFragmentAdapter.java
@@ -1,4 +1,4 @@
-package com.example.anpu.circles;
+package com.example.anpu.circles.adapter;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
@@ -11,7 +11,7 @@ import java.util.List;
  * Created by anpu on 2018/3/6.
  */
 
-class ViewPagerFragmentAdapter extends FragmentPagerAdapter {
+public class ViewPagerFragmentAdapter extends FragmentPagerAdapter {
 
     private List<Fragment> fragmentList = new ArrayList<>();
